Wed, May 29, 2024, 3:25 PM +08:00

+++ meal/meal.py
@@ -4,7 +4,7 @@ inputtime="7:50"
 
 def convert(time):
     hour, minute = time.split(":")
-    convert_minute = float(int(minute)/60)
+    convert_minute = float(float(minute)/60)
     return hour+str(convert_minute)
 #    return hour
